+++ src/index.js
@@ -23,6 +23,7 @@ export default class RNPickerSelect extends PureComponent {
                 value: PropTypes.any.isRequired,
                 key: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
                 color: ColorPropType,
+                displayValue: PropTypes.bool,
             })
         ).isRequired,
         value: PropTypes.any, // eslint-disable-line react/forbid-prop-types
@@ -36,6 +37,7 @@ export default class RNPickerSelect extends PureComponent {
         itemKey: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
         style: PropTypes.shape({}),
         children: PropTypes.any, // eslint-disable-line react/forbid-prop-types
+        onOpen: PropTypes.func,
         useNativeAndroidPickerStyle: PropTypes.bool,
 
         // Custom Modal props (iOS only)
@@ -43,7 +45,6 @@ export default class RNPickerSelect extends PureComponent {
         onDonePress: PropTypes.func,
         onUpArrow: PropTypes.func,
         onDownArrow: PropTypes.func,
-        onOpen: PropTypes.func,
         onClose: PropTypes.func,
 
         // Modal props (iOS only)
@@ -380,9 +381,10 @@ export default class RNPickerSelect extends PureComponent {
                         Platform.OS === 'ios' ? style.inputIOS : style.inputAndroid,
                         this.getPlaceholderStyle(),
                     ]}
+                    testID="text_input"
                     {...textInputProps}
                 >
-                    {selectedItem.label}
+                    {selectedItem.displayValue ? selectedItem.value : selectedItem.label}
                 </Text>
                 {this.renderIcon()}
             </View>
@@ -442,27 +444,29 @@ export default class RNPickerSelect extends PureComponent {
     }
 
     renderAndroidHeadless() {
-        const { disabled, Icon, style, pickerProps } = this.props;
+        const { disabled, Icon, style, pickerProps, onOpen } = this.props;
         const { selectedItem } = this.state;
 
         return (
-            <View style={style.headlessAndroidContainer}>
-                {this.renderTextInputOrChildren()}
-                <Picker
-                    style={[
-                        Icon ? { backgroundColor: 'transparent' } : {}, // to hide native icon
-                        defaultStyles.headlessAndroidPicker,
-                        style.headlessAndroidPicker,
-                    ]}
-                    testID="android_picker_headless"
-                    enabled={!disabled}
-                    onValueChange={this.onValueChange}
-                    selectedValue={selectedItem.value}
-                    {...pickerProps}
-                >
-                    {this.renderPickerItems()}
-                </Picker>
-            </View>
+            <TouchableWithoutFeedback onPress={onOpen} testID="android_touchable_wrapper">
+                <View style={style.headlessAndroidContainer}>
+                    {this.renderTextInputOrChildren()}
+                    <Picker
+                        style={[
+                            Icon ? { backgroundColor: 'transparent' } : {}, // to hide native icon
+                            defaultStyles.headlessAndroidPicker,
+                            style.headlessAndroidPicker,
+                        ]}
+                        testID="android_picker_headless"
+                        enabled={!disabled}
+                        onValueChange={this.onValueChange}
+                        selectedValue={selectedItem.value}
+                        {...pickerProps}
+                    >
+                        {this.renderPickerItems()}
+                    </Picker>
+                </View>
+            </TouchableWithoutFeedback>
         );
     }